Optimize gamma_delta calculation
First, we're now storing the whole Vbar array in memory so that we
can access the elements faster. This gives ~10x speed increase over
calculating the elements using the stored values for the 1D matrix
and doing some index magic.

In addition to that, we get ~4x increase by reorganizing the loop in
gamma_delta. We make sure that we iterate over columns in the Vbar
matrix, but we also minimize the number of loops we do by just looping
over (i<j, k<l) values.